Update AABBCollisionDetection.h
remembered that I forgot to add something vital to vec3 or physicsobject. I am pretty sure I didn't add a way to initialize data for soa. I will be working on that and ebugging most of next week as I am current ahead of schedule for adding items. Thankfully, the code is pretty straightforward so I will be able to test these bottom up to confirm they work then work out the kinks. If I saw it evolving to be more complicated at this stage, I would have stopped the updates and focus on debugging. Its easier to work on once you know the base is strong sort of idea.