Implement broad phase collision detection structures
Added structures and methods for broad phase collision detection management. Did mostly research on this today and some work into a Machine Learning experiment. The work mostly was to jot down the notes and ideas of a mechanism I would love to try to navigate. That will probably be my next project after this engine.